Reorganize some APIs
